Stream management and server-side events
Implements a gRPC streaming server where events are generated on the
server and broadcast to all connected clients. Clients respond to the
events with reports that the server handles. The server manages streams
ensuring that it can close streams if needed without waiting for the
client to send a message.